fix: add .gitattributes to resolve UTF-16 encoding issues
HTML and XML group policy reports are UTF-16 little endian encoded -- which is problematic when storing the files in Git using the default configuration (in particular when using `git diff` to compare files). To avoid these encoding issues, assign attributes to *.html and *.xml files to force Git to handle them as expected.